+++ create-disk-copy-and-add-to-vm.sh
@@ -662,7 +662,7 @@ verify_copy() {
 
 setup() {
     define_colours
-    PROJECT_VERSION="3.2.0"; SCRIPT_VERSION="3.2.0"
+    PROJECT_VERSION="3.2.1"; SCRIPT_VERSION="3.2.1"
     MODE="hot"; MODE_ARG=""
     ARG_COUNT=0; ARG1=""; ARG2=""; ARG3=""; ARG4=""; ARG5=""
     SOURCE_FORM=""; SOURCE_INPUT=""; SOURCE_VM_INPUT=""; SOURCE_SELECTOR=""